requirements.txt
Adafruit-Blinka==8.19.0
adafruit-circuitpython-busdevice==5.2.6
adafruit-circuitpython-neopixel==6.3.9
adafruit-circuitpython-pixelbuf==2.0.2
adafruit-circuitpython-requests==1.14.0
adafruit-circuitpython-typing==1.9.4
Adafruit-PlatformDetect==3.46.0
Adafruit-PureIO==1.1.11
aiohttp==3.8.4
aiohttp-retry==2.8.3
aiosignal==1.3.1
arandr==0.1.10
astroid==2.5.1
asttokens==2.0.4
async-timeout==4.0.2
attrs==23.1.0
automationhat==0.2.0
beautifulsoup4==4.9.3
blinker==1.4
blinkt==0.1.2
board==1.0
buttonshim==0.0.2
CacheControl==0.12.11
cachetools==5.3.0
Cap1xxx==0.1.3
certifi==2020.6.20
cffi==1.15.1
chardet==4.0.0
charset-normalizer==3.1.0
click==7.1.2
colorama==0.4.4
colorzero==1.1
contourpy==1.0.6
cryptography==39.0.1
cupshelpers==1.0
cycler==0.11.0
dbus-python==1.2.16
distro==1.5.0
docutils==0.16
drumhat==0.1.0
envirophat==1.0.0
ExplorerHAT==0.4.2
firebase-admin==6.1.0
Flask==1.1.2
fonttools==4.38.0
fourletterphat==0.1.0
frozenlist==1.3.3
google-api-core==2.11.0
google-api-python-client==2.79.0
google-auth==2.16.1
google-auth-httplib2==0.1.0
google-cloud-core==2.3.2
google-cloud-firestore==2.9.1
google-cloud-storage==2.7.0
google-crc32c==1.5.0
google-resumable-media==2.4.1
googleapis-common-protos==1.58.0
gpio==0.3.0
gpiozero==1.6.2
grpcio==1.51.1
grpcio-status==1.51.1
html5lib==1.1
httplib2==0.21.0
idna==2.10
igraph==0.10.3
importlib-resources==5.12.0
isort==5.6.4
itsdangerous==1.1.0
jedi==0.18.0
Jinja2==2.11.3
kiwisolver==1.4.4
lazy-object-proxy==0.0.0
logilab-common==1.8.1
lxml==4.6.3
MarkupSafe==1.1.1
matplotlib==3.6.2
mccabe==0.6.1
microdotphat==0.2.1
mote==0.0.4
motephat==0.0.3
msgpack==1.0.4
multidict==6.0.4
mypy==0.812
mypy-extensions==0.4.3
numpy==1.24.1
oauthlib==3.1.0
opencv-python==4.5.3.56
packaging==22.0
pantilthat==0.0.7
parso==0.8.1
pexpect==4.8.0
pgzero==1.2
phatbeat==0.1.1
pianohat==0.1.0
picamera==1.13
picamera2==0.3.9
pidng==4.0.9
piexif==1.1.3
piglow==1.2.5
pigpio==1.78
Pillow==8.1.2
proto-plus==1.22.2
protobuf==4.22.0
psutil==5.8.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycairo==1.16.2
pycparser==2.21
pycups==2.0.1
pyftdi==0.54.0
pygame==1.9.6
Pygments==2.7.1
PyGObject==3.38.0
pyinotify==0.9.6
PyJWT==2.6.0
pylint==2.7.2
PyOpenGL==3.1.5
pyOpenSSL==20.0.1
pyparsing==3.0.9
PyQt5==5.15.2
PyQt5-sip==12.8.1
pyserial==3.5
pysmbc==1.0.23
python-apt==2.2.1
python-dateutil==2.8.2
python-prctl==1.7
pytz==2023.3
pyusb==1.2.1
rainbowhat==0.1.0
reportlab==3.5.59
requests==2.25.1
requests-oauthlib==1.0.0
responses==0.12.1
roman==2.0.0
rpi-ws281x==5.0.0
RPi.GPIO==0.7.1
rsa==4.9
RTIMULib==7.2.1
SCons==4.0.1
scrollphat==0.0.7
scrollphathd==1.2.1
Send2Trash==1.6.0b1
sense-hat==2.4.0
simplejpeg==1.6.4
simplejson==3.17.2
six==1.16.0
skywriter==0.0.7
sn3218==1.2.7
soupsieve==2.2.1
spidev==3.5
ssh-import-id==5.10
sysv-ipc==1.1.0
texttable==1.6.7
thonny==4.0.1
toml==0.10.1
touchphat==0.0.1
twilio==8.2.1
twython==3.8.2
typed-ast==1.4.2
typing_extensions==4.6.3
unicornhathd==0.0.4
uritemplate==4.1.1
urllib3==1.26.5
v4l2-python3==0.3.2
webencodings==0.5.1
Werkzeug==1.0.1
wrapt==1.12.1
yarl==1.9.2
zipp==3.15.0
